electronbackend/requirements.txt
flask>=3.0.2
flask-socketio>=5.3.6
flask-cors>=4.0.0
python-dotenv>=1.0.1
websocket-client>=1.7.0
requests>=2.31.0
watchdog>=4.0.0
google-cloud-aiplatform==1.95.1
google-generativeai>=0.5.0
numpy==2.2.6
sounddevice==0.5.2 
Pillow>=10.0.0 
openai>=1.0.0 
PyJWT>=2.8.0
bcrypt>=4.0.0
sqlite3 